compile ok

+++ demo/Server.cpp
@@ -748,7 +748,7 @@ void Server::initWorker(int consumerFd,int producerFd,int payloadConsumerFd,int 
 json Server::getStringFromBase64(std::string payload)
 {
     Base64 base;
-    auto dec = base.Decode(payload.c_str(),payload.length());
+    auto dec = base.Decode(payload.c_str(),(int)payload.length());
     return json::parse(dec);
 }
